pyproject.toml
[project]
name = "candor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "convokit>=3.0.1",
    "ipykernel>=6.31.0",
    "pandas>=2.3.3",
    "plotnine>=0.13.6",
    "pyannote-audio==3.4.0",
    "python-dotenv>=1.1.1",
    "torchaudio==2.8.0",
]
